Fixes example not being rendered in safe area of screen
Also cleaned up some linter related issues making it hard to work with
the files
